<refactor> modify：项目大重构  服务端每一帧都发送角色状态同步帧

+++ Assets/Scripts/Client/Bussiness/BattleBussiness/Repo/BulletRepo.cs
@@ -18,7 +18,7 @@ namespace Game.Client.Bussiness.BattleBussiness.Repo
             bulletList = new List<BulletEntity>();
         }
 
-        public BulletEntity GetByBulletId(ushort bulletId)
+        public BulletEntity GetByBulletId(int bulletId)
         {
             return bulletList.Find((entity) => entity.IDComponent.EntityId == bulletId);
         }
